% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQuantileSplit.R
\name{plotQuantileSplit}
\alias{plotQuantileSplit}
\title{Kaplan Meijer plot for continuous variable.}
\usage{
plotQuantileSplit(in.df, survCol, sigCol, nrsplits = 4, ...)
}
\arguments{
\item{in.df}{Input dataframe}

\item{survCol}{Column containing the survival object generated by Surv(). Can be either the name as a string or the column number.}

\item{sigCol}{Column containing the continuous variable. Can be either the name as a string or the column number.}

\item{nrsplits}{Numeric. Into how many quantiles should the variable be split?}

\item{...}{Additional arguments to plotSurvCategory}
}
\description{
`plotQuantileSplit()` splits a patient group into n subgroups and generates an annotated KM-plot. It shows either the univariate HR (two groups) or logtest p-value (more groups).
}
