% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurvCategory.R
\name{plotSurvCategory}
\alias{plotSurvCategory}
\title{Kaplan Meijer plot for patient subgroups.}
\usage{
plotSurvCategory(in.df, survCol, sigCol, title = NA, ymin = 0,
  showlegend = T, colours = NA, linetypes = 1)
}
\arguments{
\item{in.df}{Input dataframe}

\item{survCol}{Column containing the survival object generated by Surv(). Can be either the name as a string or the column number.}

\item{sigCol}{Column containing the patient groups. Can be either the name as a string or the column number.}

\item{title}{Character. Plot title. Defaults to the name of the "sigCol" column}

\item{ymin}{Numeric. The lowerbound value on the y-axis. Use this to emphasize differences.}

\item{colours}{Character vector of colours to use for the different patient groups.}

\item{linetypes}{Character vector of linetypes to use for the different patient groups.}

\item{showLegend}{Boolean. Add a legend to the plot?}
}
\description{
`plotSurvCategory()` generates an annotated KM-plot and shows either the univariate HR (two groups) or logtest p-value (more groups).
}
